// webapps/backend/src/api-client/api.ts
export interface LoginDto {
    /**
     * The username of the user
     */
    'username': string;
    /**
     * The password of the user
     */
    'password': string;
}
export interface MarketSignal {
    'id': number;
    'symbol': string;
    'last_close': number;
    'signal': string;
    'trend': string;
    'rsi_14': number;
    'rsi_overbought': number;
    'rsi_oversold': number;
    'ema_50': number;
    'ema_200': number;
    'golden_cross': boolean;
    'death_cross': boolean;
    'updated_at': string;
}
export interface RefreshDto {
    'refreshToken': string;
}
// ...
export class AuthenticationApi extends BaseAPI {
    /**
     * 
     * @param {LoginDto} loginDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public authControllerLogin(loginDto: LoginDto, options?: RawAxiosRequestConfig) {
        return AuthenticationApiFp(this.configuration).authControllerLogin(loginDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public authControllerLogout(options?: RawAxiosRequestConfig) {
        return AuthenticationApiFp(this.configuration).authControllerLogout(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public authControllerLogoutAll(options?: RawAxiosRequestConfig) {
        return AuthenticationApiFp(this.configuration).authControllerLogoutAll(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {RefreshDto} refreshDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public authControllerRefresh(refreshDto: RefreshDto, options?: RawAxiosRequestConfig) {
        return AuthenticationApiFp(this.configuration).authControllerRefresh(refreshDto, options).then((request) => request(this.axios, this.basePath));
    }
}